# Simplify the block layout.