--- a/src/lsm/binary_search_benchmark.zig
+++ b/src/lsm/binary_search_benchmark.zig
@@ -4,34 +4,20 @@ const math = std.math;
 
 const binary_search_keys_upsert_index = @import("./binary_search.zig").binary_search_keys_upsert_index;
 const binary_search_values_upsert_index = @import("./binary_search.zig").binary_search_values_upsert_index;
-const eytzinger = @import("./eytzinger.zig").eytzinger;
 
 const GiB = 1 << 30;
 const searches = 500_000;
 
 const kv_types = .{
-    .{ .key_size = 8, .value_size = 128 },
-    .{ .key_size = 8, .value_size = 64 },
-    .{ .key_size = 16, .value_size = 16 },
-    .{ .key_size = 32, .value_size = 32 },
+    .{ .key_size = @sizeOf(u64), .value_size = 128 },
+    .{ .key_size = @sizeOf(u64), .value_size = 64 },
+    .{ .key_size = @sizeOf(u128), .value_size = 16 },
+    .{ .key_size = @sizeOf(u256), .value_size = 32 },
 };
 
-// keys_per_summary = values_per_page / summary_fraction
-const summary_fractions = .{ 4, 8, 16, 32 };
-const values_per_page = .{ 128, 256, 512, 1024, 2048, 4096, 8192 };
-const body_fmt = "{:_>2}B/{:_>3}B {:_>4}/{:_>4} {s}{s}: WT={:_>6}ns UT={:_>6}ns" ++
+const values_per_page = .{ 128, 256, 512, 1024, 2 * 1024, 4 * 1024, 8 * 1024 };
+const body_fmt = "K={:_>2}B V={:_>3}B N={:_>4} {s}{s}: WT={:_>6}ns UT={:_>6}ns" ++
     " CY={:_>6} IN={:_>6} CR={:_>5} CM={:_>5} BM={}\n";
-
-const summary_sizes = blk: {
-    var sizes: [values_per_page.len][summary_fractions.len]usize = undefined;
-    for (values_per_page, 0..) |values_count, v| {
-        for (summary_fractions, 0..) |fraction, k| {
-            // Set in reverse order so that the summary sizes ascend.
-            sizes[v][summary_fractions.len - k - 1] = values_count / fraction;
-        }
-    }
-    break :blk sizes;
-};
 
 pub fn main() !void {
     std.log.info("Samples: {}", .{searches});
@@ -56,28 +42,23 @@ pub fn main() !void {
     var blob = try arena.allocator().alloc(u8, blob_size);
 
     inline for (kv_types) |kv| {
-        inline for (values_per_page, 0..) |values_count, v| {
-            inline for (summary_sizes[v]) |keys_count| {
-                try run_benchmark(.{
-                    .blob_size = blob_size,
-                    .key_size = kv.key_size,
-                    .value_size = kv.value_size,
-                    .keys_count = keys_count,
-                    .values_count = values_count,
-                    .searches = searches,
-                }, blob, prng.random());
-            }
+        inline for (values_per_page) |values_count| {
+            try run_benchmark(.{
+                .blob_size = blob_size,
+                .key_size = kv.key_size,
+                .value_size = kv.value_size,
+                .values_count = values_count,
+                .searches = searches,
+            }, blob, prng.random());
         }
     }
 }
 
 fn run_benchmark(comptime layout: Layout, blob: []u8, random: std.rand.Random) !void {
     assert(blob.len == layout.blob_size);
-    const Eytzinger = eytzinger(layout.keys_count - 1, layout.values_count);
     const V = Value(layout);
     const K = V.Key;
     const Page = struct {
-        keys: [layout.keys_count]K,
         values: [layout.values_count]V,
     };
     const page_count = layout.blob_size / @sizeOf(Page);
@@ -92,107 +73,51 @@ fn run_benchmark(comptime layout: Layout, blob: []u8, random: std.rand.Random) !
     random.bytes(std.mem.sliceAsBytes(pages));
     for (pages) |*page| {
         for (&page.values, 0..) |*value, i| value.key = i;
-        Eytzinger.layout_from_keys_or_values(K, V, V.key_from_value, V.max_key, &page.values, &page.keys);
     }
 
     const stdout = std.io.getStdOut().writer();
-    inline for (&.{ true, false }) |prefetch| {
-        var benchmark = try Benchmark.begin();
-        var i: usize = 0;
-        var v: usize = 0;
-        while (i < layout.searches) : (i += 1) {
-            const page_index = page_picker[i % page_picker.len];
-            const target = value_picker[v % value_picker.len];
-            const page = &pages[page_index];
-            const bounds = Eytzinger.search_values(
+    var benchmark = try Benchmark.begin();
+    var i: usize = 0;
+    var v: usize = 0;
+    while (i < layout.searches) : (i += 1) {
+        const target = value_picker[v % value_picker.len];
+        const page = &pages[page_picker[i % page_picker.len]];
+        const hit = page.values[
+            binary_search_values_upsert_index(
                 K,
                 V,
+                V.key_from_value,
                 V.key_compare,
-                prefetch,
-                &page.keys,
-                &page.values,
+                page.values[0..],
                 target,
-            );
-            const hit = bounds[
-                binary_search_values_upsert_index(
-                    K,
-                    V,
-                    V.key_from_value,
-                    V.key_compare,
-                    bounds,
-                    target,
-                    .{},
-                )
-            ];
+                .{},
+            )
+        ];
 
-            assert(hit.key == target);
-            if (i % pages.len == 0) v += 1;
-        }
-
-        const result = try benchmark.end(layout.searches);
-        try stdout.print(body_fmt, .{
-            layout.key_size,
-            layout.value_size,
-            layout.keys_count,
-            layout.values_count,
-            if (prefetch) "P" else "_",
-            "E",
-            result.wall_time,
-            result.utime,
-            result.cpu_cycles,
-            result.instructions,
-            result.cache_references,
-            result.cache_misses,
-            result.branch_misses,
-        });
+        assert(hit.key == target);
+        if (i % pages.len == 0) v += 1;
     }
-
-    {
-        var benchmark = try Benchmark.begin();
-        var i: usize = 0;
-        var v: usize = 0;
-        while (i < layout.searches) : (i += 1) {
-            const target = value_picker[v % value_picker.len];
-            const page = &pages[page_picker[i % page_picker.len]];
-            const hit = page.values[
-                binary_search_values_upsert_index(
-                    K,
-                    V,
-                    V.key_from_value,
-                    V.key_compare,
-                    page.values[0..],
-                    target,
-                    .{},
-                )
-            ];
-
-            assert(hit.key == target);
-            if (i % pages.len == 0) v += 1;
-        }
-        const result = try benchmark.end(layout.searches);
-        try stdout.print(body_fmt, .{
-            layout.key_size,
-            layout.value_size,
-            layout.keys_count,
-            layout.values_count,
-            "_",
-            "B",
-            result.wall_time,
-            result.utime,
-            result.cpu_cycles,
-            result.instructions,
-            result.cache_references,
-            result.cache_misses,
-            result.branch_misses,
-        });
-    }
+    const result = try benchmark.end(layout.searches);
+    try stdout.print(body_fmt, .{
+        layout.key_size,
+        layout.value_size,
+        layout.values_count,
+        "_",
+        "B",
+        result.wall_time,
+        result.utime,
+        result.cpu_cycles,
+        result.instructions,
+        result.cache_references,
+        result.cache_misses,
+        result.branch_misses,
+    });
 }
 
 const Layout = struct {
     blob_size: usize, // bytes allocated for all pages
     key_size: usize, // bytes per key
     value_size: usize, // bytes per value
-    keys_count: usize, // keys per page (in the summary)
     values_count: usize, // values per page
     searches: usize,
 };
@@ -314,25 +239,4 @@ fn readPerfFd(fd: std.os.fd_t) !usize {
     assert(n == @sizeOf(usize));
 
     return result;
-}
-
-fn binary_search_keys(
-    comptime layout: Layout,
-    comptime Key: type,
-    comptime V: type,
-    comptime compare_keys: fn (Key, Key) math.Order,
-    keys: []const Key,
-    values: []const V,
-    key: Key,
-) []const V {
-    assert(keys.len == layout.keys_count);
-    assert(values.len == layout.values_count);
-
-    const key_index = binary_search_keys_upsert_index(Key, compare_keys, keys, key, .{});
-    const key_stride = layout.values_count / layout.keys_count;
-    const high = key_index * key_stride;
-    if (key_index < keys.len and keys[key_index] == key) {
-        return if (high == 0) values[0..1] else values[high - 1 .. high];
-    }
-    return values[high - key_stride .. high];
 }