--- a/src/unit_tests.zig
+++ b/src/unit_tests.zig
@@ -19,7 +19,6 @@ test {
 
     _ = @import("lsm/binary_search.zig");
     _ = @import("lsm/bloom_filter.zig");
-    _ = @import("lsm/eytzinger.zig");
     _ = @import("lsm/forest.zig");
     _ = @import("lsm/groove.zig");
     _ = @import("lsm/k_way_merge.zig");