- Deletes `eytzinger.zig` and removes the layout block, along with `key_layout*` constants and derived configurations.

- Only computes bloom filters where `Table.usage == .general`, although still keeping the filter block on the layout. Asserts that filter blocks are only used for `Table.usage == .general`.

- Repurposes the `eytzinger_benchmark` as a `binary_search_benchmark`.

- Refactors the `// TODO u16 + padding.` at `schema.zig` for `TableData.Parent`.


And a surprisingly (and consistently) good benchmark result:

Before:

```
Benchmarking...
steps [2/4] zig build-exe benchmark ReleaseSafe native-linux... LLVM Emit Object... info: Benchmark running against { 127.0.0.1:48005 }
info(message_bus): connected to replica 0
1224 batches in 76.16 s
load offered = 1000000 tx/s
load accepted = 131301 tx/s
batch latency p00 = 0 ms
batch latency p10 = 20 ms
batch latency p20 = 23 ms
batch latency p30 = 26 ms
batch latency p40 = 27 ms
batch latency p50 = 28 ms
batch latency p60 = 29 ms
batch latency p70 = 30 ms
batch latency p80 = 31 ms
batch latency p90 = 32 ms
batch latency p100 = 3104 ms
transfer latency p00 = 0 ms
transfer latency p10 = 4115 ms
transfer latency p20 = 8164 ms
transfer latency p30 = 14102 ms
transfer latency p40 = 20223 ms
transfer latency p50 = 27869 ms
transfer latency p60 = 35884 ms
transfer latency p70 = 43705 ms
transfer latency p80 = 51981 ms
transfer latency p90 = 59220 ms
transfer latency p100 = 66178 ms
```

After:
```
Benchmarking...
steps [2/4] zig build-exe benchmark ReleaseSafe native-linux... LLVM Emit Object... info: Benchmark running against { 127.0.0.1:58059 }
info(message_bus): connected to replica 0
1224 batches in 74.40 s
load offered = 1000000 tx/s
load accepted = 134399 tx/s
batch latency p00 = 0 ms
batch latency p10 = 19 ms
batch latency p20 = 21 ms
batch latency p30 = 26 ms
batch latency p40 = 27 ms
batch latency p50 = 28 ms
batch latency p60 = 29 ms
batch latency p70 = 30 ms
batch latency p80 = 31 ms
batch latency p90 = 33 ms
batch latency p100 = 3027 ms
transfer latency p00 = 0 ms
transfer latency p10 = 4182 ms
transfer latency p20 = 7481 ms
transfer latency p30 = 12715 ms
transfer latency p40 = 19345 ms
transfer latency p50 = 26541 ms
transfer latency p60 = 34645 ms
transfer latency p70 = 42122 ms
transfer latency p80 = 50095 ms
transfer latency p90 = 57515 ms
transfer latency p100 = 64423 ms
```